[wireless] Backport iwlwifi driver from linux-4.11-rc1
Message-id: <1488820820-16439-7-git-send-email-sgruszka@redhat.com>
Patchwork-id: 170933
O-Subject: [RHEL7 PATCH 06/10] Backport iwlwifi driver from linux-4.11-rc1
Bugzilla: 1387986 1385917 1385746 1351045 1421597
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

BZ#1421597
BZ#1387986
BZ#1385917
BZ#1385746
BZ#1351045

Description:
Sync iwlwifi driver to linux-4.11-rc1, up to
commit c1ae3cfa0e89fa1a7ecc4c99031f5e9ae99d9201.

Brew:
https://brewweb.engineering.redhat.com/brew/taskinfo?taskID=126932211

Testing:
Tested on IWL 5100, 6300 and 7260 adapters.

+++ drivers/net/wireless/intel/iwlwifi/dvm/calib.c
@@ -902,11 +902,10 @@ static void iwlagn_gain_computation(struct iwl_priv *priv,
 		data->delta_gain_code[i] =
 			min(abs(delta_g),
 #if 0 /* abs() is different on RHEL */
-			(s32) CHAIN_NOISE_MAX_DELTA_GAIN_CODE);
+			CHAIN_NOISE_MAX_DELTA_GAIN_CODE);
 #else
 			(long) CHAIN_NOISE_MAX_DELTA_GAIN_CODE);
 #endif
-
 		if (delta_g < 0)
 			/*
 			 * set negative sign ...
